refactor: move all static react components to astro components
Essentially, this is changing `.tsx` files into `.astro` files,
change `className` in `class`, remove `key` attributes to html elements.

Worth noting: the use of Avatar required some change, as per
https://github.com/shadcn-ui/ui/issues/987#issuecomment-1913143431
By the way, this has the side-effect to actually load the shadcn.png image
in the testimonials (still not the case in testimonial in the hero section).
For whatever reason, the avatar was not loading previsouly
(at least in local dev).