docs: introduce this repository
This is a port of shadcn-landing-page (credits to leoMirandaa) to astro.

At this stage, this is an _exact_ port (i.e. to-the-pixel!), except for removing
mentions to shadcn/react specific stuff not applicable to this repository,
and replacing by tailwind/astro (which is the target of this repo).

Note: the font size in the header navigation is actually a bit different
between the original shadcn/react version and this version (.875rem and 17px respectively).
However, this is believed to be a bug in the original version because:
- the text is given both classes `text-[17px]` and `text-sm`
- by giving explicitly the class `text-[17px]`, we can assume this was intentionally the target font size
- for some reason, in the original version, the styling for `text-sm` takes precedence.
It is not the case in the astro version.